Refactoring files into proper folder structure
